requirements.txt
anyio==3.5.0
asgiref==3.5.1
bcrypt==3.2.2
beanie==1.10.8
certifi==2023.5.7
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.1
DateTime==5.1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.75.2
h11==0.13.0
httptools==0.4.0
idna==3.3
jose==1.0.0
lazy-model==0.0.5
motor==2.5.1
multidict==6.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
yarl==1.7.2